LPS-52484 Move to the internal package
