+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-values-factory/src/com/liferay/dynamic/data/mapping/form/values/factory/internal/DDMFormValuesFactoryImpl.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.dynamic.data.mapping.form.values.factory.impl;
+package com.liferay.dynamic.data.mapping.form.values.factory.internal;
 
 import com.liferay.dynamic.data.mapping.form.renderer.DDMFormRendererConstants;
 import com.liferay.dynamic.data.mapping.form.values.factory.DDMFormValuesFactory;
