+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-values-query/src/com/liferay/dynamic/data/mapping/form/values/query/internal/model/DDMFormFieldValueMatcher.java
@@ -12,28 +12,15 @@
  * details.
  */
 
-package com.liferay.dynamic.data.mapping.form.values.query.impl.model;
+package com.liferay.dynamic.data.mapping.form.values.query.internal.model;
 
 import com.liferay.portlet.dynamicdatamapping.storage.DDMFormFieldValue;
-import com.liferay.portlet.dynamicdatamapping.storage.DDMFormValues;
-
-import java.util.List;
 
 /**
  * @author Pablo Carvalho
  */
-public interface DDMFormValuesFilter {
+public interface DDMFormFieldValueMatcher {
 
-	public List<DDMFormFieldValue> filter(DDMFormValues ddmFormValues);
-
-	public List<DDMFormFieldValue> filter(
-		List<DDMFormFieldValue> baseDDMFormFieldValues);
-
-	public boolean isGreedy();
-
-	public void setDDMFormFieldValueMatcher(
-		DDMFormFieldValueMatcher ddmFormFieldValueMatcher);
-
-	public void setGreedy(boolean greedy);
+	public boolean matches(DDMFormFieldValue ddmFormFieldValue);
 
 }
