+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-values-query/src/com/liferay/dynamic/data/mapping/form/values/query/internal/parser/DDMFormValuesQueryLexer.java
@@ -14,7 +14,7 @@
  * details.
  */
 
-package com.liferay.dynamic.data.mapping.form.values.query.impl.parser;
+package com.liferay.dynamic.data.mapping.form.values.query.internal.parser;
 
 import org.antlr.v4.runtime.Lexer;
 import org.antlr.v4.runtime.CharStream;
